#606: Move Neow3jExpress to io.neow3j.protocol package
